complete the move
