+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/events/ApplicationEventProcessorTest.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.clients.consumer.internals.events;
 
+import org.apache.kafka.clients.consumer.OffsetAndMetadata;
 import org.apache.kafka.clients.consumer.internals.CommitRequestManager;
 import org.apache.kafka.clients.consumer.internals.ConsumerMembershipManager;
 import org.apache.kafka.clients.consumer.internals.ConsumerMetadata;
@@ -27,21 +28,34 @@ import org.apache.kafka.clients.consumer.internals.OffsetsRequestManager;
 import org.apache.kafka.clients.consumer.internals.RequestManagers;
 import org.apache.kafka.clients.consumer.internals.SubscriptionState;
 import org.apache.kafka.clients.consumer.internals.TopicMetadataRequestManager;
+import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.utils.LogContext;
+import org.apache.kafka.common.utils.MockTime;
+import org.apache.kafka.common.utils.Time;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
+import java.util.stream.Stream;
 
+import static org.apache.kafka.clients.consumer.internals.events.CompletableEvent.calculateDeadlineMs;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 public class ApplicationEventProcessorTest {
+    private final Time time = new MockTime();
     private final CommitRequestManager commitRequestManager = mock(CommitRequestManager.class);
     private final HeartbeatRequestManager heartbeatRequestManager = mock(HeartbeatRequestManager.class);
     private final ConsumerMembershipManager membershipManager = mock(ConsumerMembershipManager.class);
@@ -89,6 +103,46 @@ public class ApplicationEventProcessorTest {
         setupProcessor(false);
         processor.process(new UnsubscribeEvent(0));
         verify(subscriptionState).unsubscribe();
+    }
+
+    @ParameterizedTest
+    @MethodSource("applicationEvents")
+    public void testApplicationEventIsProcessed(ApplicationEvent e) {
+        ApplicationEventProcessor applicationEventProcessor = mock(ApplicationEventProcessor.class);
+        applicationEventProcessor.process(e);
+        verify(applicationEventProcessor).process(any(e.getClass()));
+    }
+
+    private static Stream<Arguments> applicationEvents() {
+        Map<TopicPartition, OffsetAndMetadata> offset = new HashMap<>();
+        final long currentTimeMs = 12345;
+        return Stream.of(
+                Arguments.of(new PollEvent(100)),
+                Arguments.of(new NewTopicsMetadataUpdateRequestEvent()),
+                Arguments.of(new AsyncCommitEvent(new HashMap<>())),
+                Arguments.of(new SyncCommitEvent(new HashMap<>(), 500)),
+                Arguments.of(new ResetPositionsEvent(500)),
+                Arguments.of(new ValidatePositionsEvent(500)),
+                Arguments.of(new TopicMetadataEvent("topic", Long.MAX_VALUE)),
+                Arguments.of(new AssignmentChangeEvent(offset, currentTimeMs)));
+    }
+
+    @ParameterizedTest
+    @ValueSource(booleans = {true, false})
+    public void testListOffsetsEventIsProcessed(boolean requireTimestamp) {
+        ApplicationEventProcessor applicationEventProcessor = mock(ApplicationEventProcessor.class);
+        Map<TopicPartition, Long> timestamps = Collections.singletonMap(new TopicPartition("topic1", 1), 5L);
+        ApplicationEvent e = new ListOffsetsEvent(timestamps, calculateDeadlineMs(time, 100), requireTimestamp);
+        applicationEventProcessor.process(e);
+        verify(applicationEventProcessor).process(any(ListOffsetsEvent.class));
+    }
+
+    @Test
+    public void testResetPositionsProcess() {
+        ApplicationEventProcessor applicationEventProcessor = mock(ApplicationEventProcessor.class);
+        ResetPositionsEvent event = new ResetPositionsEvent(calculateDeadlineMs(time, 100));
+        applicationEventProcessor.process(event);
+        verify(applicationEventProcessor).process(any(ResetPositionsEvent.class));
     }
 
     private List<NetworkClientDelegate.UnsentRequest> mockCommitResults() {
